chore: commit sprint plan execution changes
Stage and commit the remaining sprint-plan implementation work, including governance gates, compatibility shim checks, tenant/auth hardening tests, frontend routing cleanup, Layer 4 canonical import cleanup, and targeted contract/security regressions.

Validation:

- git diff --cached --check

- prior sprint-plan gate suite completed before commit

+++ services/layer2-extraction/src/layer2_extraction/api/retry_queue.py
@@ -32,7 +32,7 @@ class PendingRecordLike(Protocol):
 
 
 def retry_delay_seconds(*, retry_base_seconds: int, retry_count: int) -> int:
-    return retry_base_seconds * max(1, 2 ** max(retry_count - 1, 0))
+    return int(retry_base_seconds * max(1, 2 ** max(retry_count - 1, 0)))
 
 
 def next_retry_at(
